requirements.txt
pybit>=2.6.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
ta-lib>=0.4.0  # Для технических индикаторов